feat: Python and Rust consumers share common cluster resolution code
Centralize the code for validating and resolving the Kafka and ClickHouse
clusters and other consumer config to be passed to the Python and Rust consumers.
Will likely also be shared by the DLQ consumer and multistorage consumer.
